requirements.txt
Beaker==1.7.0
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
argparse==1.2.1
bottle==0.12.8
py-bcrypt==0.4
simplejson==3.8.0
wsgiref==0.1.2
